feat: more scatter simplifications for add/sub/div
